Update LastDefender.js
lastDefender version 0.2 is finally out !  new enemy added! the "Talpa"!
the talpa spawns underground  in a random position but will not spawn under the hut, when the talpa head is not under the ground the talpa pops out of it and come for your hut.
the ground id now more up then before because of the talpa.
resolved the bug that would make the bullet have more speed with multiple enemy on the board.